+++ src/pages/page2/page2.ts
@@ -3,7 +3,7 @@
 import { Component }        from '@angular/core';
 
 @Component({
-  templateUrl: 'page2.html',
+  templateUrl: './page2.html',
 })
 
 export class Page2 {